+++ comics/comics/gregcomic.py
@@ -7,21 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://gregcomic.com/'
     start_date = '2011-06-01'
+    end_date = '2013-12-20'
+    active = False
     rights = 'Chur Yin Wan'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 180
-    time_zone = 'US/Eastern'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://gregcomic.com/feed/')
-        for entry in feed.for_date(pub_date):
-            if 'Comics' not in entry.tags:
-                continue
-            title = entry.title
-            url = entry.summary.src('img[src*="/comics-rss/"]')
-            if not url:
-                continue
-            url = url.replace('-rss', '')
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published